Move the linker to its own file
